+++ app/app.module.ts
@@ -8,6 +8,8 @@ import { AboutComponent }      from './components/about/about.component';
 import { ProfileComponent }      from './components/profile/profile.component';
 import { ContactComponent }      from './components/contact/contact.component';
 
+import { PageComponent }      from './components/page/page.component';
+import { PageNotFoundComponent }      from './components/404/404.component';
 import {routing} from './app.routing';
 
 
@@ -30,7 +32,7 @@ import { FormsModule }   from '@angular/forms';
       TranslateModule.forRoot(),
       FormsModule
    ],
-  declarations: [ AppComponent, HomeComponent, AboutComponent, ProfileComponent, ContactComponent, StarRatingPipe, SanitizeHTML ],
+  declarations: [ AppComponent, HomeComponent, AboutComponent, ProfileComponent, ContactComponent, PageComponent, PageNotFoundComponent, StarRatingPipe, SanitizeHTML ],
   bootstrap:    [ AppComponent  ]
 })
 export class AppModule { }